IGNITE-24454 Cache partition files calculation in NodeFileTree
