+++ modules/core/src/main/java/org/apache/ignite/internal/processors/cache/persistence/filename/SnapshotFileTree.java
@@ -20,12 +20,28 @@ package org.apache.ignite.internal.processors.cache.persistence.filename;
 import java.io.File;
 import org.jetbrains.annotations.Nullable;
 
+import static org.apache.ignite.internal.pagemem.PageIdAllocator.INDEX_PARTITION;
+import static org.apache.ignite.internal.pagemem.PageIdAllocator.MAX_PARTITION_ID;
+
 /**
  * {@link NodeFileTree} extension with the methods required to work with snapshot file tree.
  * During creation, full snapshot, creates the same file tree as regular node.
  * But, using snapshot directory as root.
  */
 public class SnapshotFileTree extends NodeFileTree {
+    /** File with delta pages suffix. */
+    public static final String DELTA_SUFFIX = ".delta";
+
+    /** File with delta pages index suffix. */
+    public static final String DELTA_IDX_SUFFIX = ".idx";
+
+
+    /** File name template consists of delta pages. */
+    public static final String PART_DELTA_TEMPLATE = PART_FILE_TEMPLATE + DELTA_SUFFIX;
+
+    /** File name template for index delta pages. */
+    public static final String INDEX_DELTA_NAME = INDEX_FILE_NAME + DELTA_SUFFIX;
+
     /** Snapshot name. */
     private final String name;
 
@@ -62,9 +78,28 @@ public class SnapshotFileTree extends NodeFileTree {
     }
 
     /**
+     * @param cacheDirName Cache dir name.
+     * @param partId Cache partition identifier.
+     * @return A file representation.
+     */
+    public File partDeltaFile(String cacheDirName, int partId) {
+        return new File(tmpFt.cacheStorage(cacheDirName), partDeltaFileName(partId));
+    }
+
+    /**
+     * @param partId Partition id.
+     * @return File name of delta partition pages.
+     */
+    public static String partDeltaFileName(int partId) {
+        assert partId <= MAX_PARTITION_ID || partId == INDEX_PARTITION;
+
+        return partId == INDEX_PARTITION ? INDEX_DELTA_NAME : String.format(PART_DELTA_TEMPLATE, partId);
+    }
+
+    /**
      * @return Path to the snapshot root directory.
      */
-    public static File root(NodeFileTree ft, String name, String path) {
+    private static File root(NodeFileTree ft, String name, String path) {
         assert name != null : "Snapshot name cannot be empty or null.";
 
         return path == null ? new File(ft.snapshotsRoot(), name) : new File(path, name);
